requirements.txt
streamlit==1.40.0
groq>=0.12.0
httpx<0.28.0
python-dotenv==1.0.1

